Moved interface Color into a folder Game
